// src/modules/employees/Employee.interface.tsx
export interface IEmployees {
    id: string
    primerApellido: string
    segundoApellido: string
    primerNombre: string
    otrosNombres?: string
    pais: string
    tIdentificacion: string
    nIdentificacion: string
    correo: string
    fechaIngreso: string
    area: string
    estado: string
    fechaCreacion: string
    fechaActualizacion?: string
}